% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{add_manta}
\alias{add_manta}
\title{Add Manta output to somatic object}
\usage{
add_manta(somatic, fn_manta, MIN_READS = 2)
}
\arguments{
\item{somatic}{incoming somatic object to modify, writes n_bnd, n_del, n_dup, n_ins, n_inv, manta
where manta is the vcf file read by VariantAnnotation}

\item{fn_manta}{manta output file}

\item{MIN_READS}{discard Manta output with fewer than this number of reads in alternate}
}
\description{
Add Manta output to somatic object
}
