% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{add_SNV_mutect}
\alias{add_SNV_mutect}
\title{Add Mutect results to somatic object}
\usage{
add_SNV_mutect(somatic, fn_mutect_clinvar_snv)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{fn_mutect_clinvar_snv}{path to SNVs}
}
\description{
Reads files that have been processed by snpSift extractFields in the annotate_mutation_calls
step of the somatic workflow. Currently only handles Mutect version 1.
}
