% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{add_SNV_strelka}
\alias{add_SNV_strelka}
\title{Add Strelka results to somatic object}
\usage{
add_SNV_strelka(somatic, fn_strelka_clinvar_snv, fn_strelka_clinvar_indel)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{fn_strelka_clinvar_snv}{path to SNVs}

\item{fn_strelka_clinvar_indel}{path to indels}
}
\description{
Reads files that have been processed by snpSift extractFields in the annotate_mutation_calls
step of the somatic workflow
}
