% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{add_SNV_strelka_germline}
\alias{add_SNV_strelka_germline}
\title{add_SNV_strelka_germline}
\usage{
add_SNV_strelka_germline(somatic, fn_strelka_germline)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{fn_strelka_germline}{path to germline VCF file}
}
\description{
Deprecated; strelka germline calling does not appear reliable in our hands
}
