% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_fusions.R
\name{add_SV_gridss}
\alias{add_SV_gridss}
\title{Wrapper function to add structural variants and fusions annotations from Gridss and Linx, respectively.}
\usage{
add_SV_gridss(somatic, fn_gripss, MIN_READS = 2)
}
\arguments{
\item{somatic}{somatic data object to modify}

\item{fn_gripss}{path to structural variants filtered gridss vcf file (e.g. SAMPLE_ID_GRIPSS.somatic.filtered.vcf)}

\item{MIN_READS}{minimum number of reads to consider}
}
\description{
Wrapper function to add structural variants and fusions annotations from Gridss and Linx, respectively.
}
