% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{adjust_CNA_segments_filter}
\alias{adjust_CNA_segments_filter}
\title{Filter segments for overlap with centromere and minimum size to cut down noise}
\usage{
adjust_CNA_segments_filter(
  genome,
  somatic,
  max_percent_centromere_overlap = 0.2,
  min_segment_percent = 5e-05
)
}
\arguments{
\item{genome}{genome object created by Genome(), contains centromere locations}

\item{somatic}{somatic object containing segments}

\item{max_percent_centromere_overlap}{eliminate segments where total overlap exceeds this percent, 0-1}

\item{min_segment_percent}{minimal size of a segment expressed as a percent of the chromosome size}
}
\description{
Filter segments for overlap with centromere and minimum size to cut down noise
}
