% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{calculate_CNA_matrix}
\alias{calculate_CNA_matrix}
\title{Given a bin size, converts genome-wide segments into values per bin}
\usage{
calculate_CNA_matrix(SO, genome, bin_size = 3e+06)
}
\arguments{
\item{SO}{list of somatic objects, final matrix will include all}

\item{genome}{object created by Genome(), used for chromosome sizes}

\item{bin_size}{size of each bin, 3000000 creates approximately 1000 bins}
}
\description{
Returns a matrix with one row per bin and one column per somatic object in SO
}
