% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{calculate_CNA_segments_clonality_CLONET}
\alias{calculate_CNA_segments_clonality_CLONET}
\title{correct raw segment CNA values using calculated purity & ploidy}
\usage{
calculate_CNA_segments_clonality_CLONET(
  genome,
  somatic,
  pileup_normal,
  pileup_tumor,
  sex = "M",
  segment_source = "copycat",
  experiment_type = "WGS",
  min_required_SNPs = 10,
  n_cores = 1,
  purity_override = purity_override,
  ploidy_override = ploidy_override
)
}
\arguments{
\item{genome}{object created by call to Genome()}

\item{somatic}{somatic data object to modify}

\item{pileup_normal}{pileup file created by ASEQ program}

\item{pileup_tumor}{pileup file created by ASEQ program}

\item{sex}{sample sex, M or F}

\item{segment_source}{source, default copycat}

\item{experiment_type}{one of WGS, WES}

\item{min_required_SNPs}{minimum number of SNPs in a segment to make a call}

\item{n_cores}{allow for multithreaded calculations}

\item{purity_override}{use this value instead of estimating purity}

\item{ploidy_override}{use this value instead of estimating ploidy}
}
\description{
CLONET's estimate can fail, so this code has some built-in attempts to retry
the calculation with parameters chosen to increase the chances of a successful
convergence. This is not an ideal solution.
}
\details{
If purity and ploidy cannot be estimated but you still want to correct the segment
copy number estimates, pass in purity_override and/or ploidy_override parameters
}
