% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{calculate_genome_bins}
\alias{calculate_genome_bins}
\title{used to bin the genome into bins of size bin_size}
\usage{
calculate_genome_bins(genome, bin_size)
}
\arguments{
\item{genome}{object created by Genome, source of chromosome sizes}

\item{bin_size}{nucleotides per bin}
}
\description{
the last bin may be smaller than bin_size
returns data frame: chrom, start, end
}
