% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_fusions.R
\name{calculate_gridss_summary}
\alias{calculate_gridss_summary}
\title{Summarize the results of GRIDSS}
\usage{
calculate_gridss_summary(somatic)
}
\arguments{
\item{somatic}{somatic data object to modify}
}
\description{
Summarize the results of GRIDSS
}
