% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{calculate_manta_summary}
\alias{calculate_manta_summary}
\title{Collate manta data into list_sv, including start/stop of each variant type
and the evidence for microhomology at deletions}
\usage{
calculate_manta_summary(somatic, nt_for_microhomology = 15)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{nt_for_microhomology}{number of nucleotides to search on shoulder of deletions}
}
\description{
Collate manta data into list_sv, including start/stop of each variant type
and the evidence for microhomology at deletions
}
