% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{calculate_SNV_signatures}
\alias{calculate_SNV_signatures}
\title{Calculate mutation signatures with ReconstructSigs methods}
\usage{
calculate_SNV_signatures(somatic, tri.counts.method = "genome", min_depth = 10)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{tri.counts.method}{passed through to whichSignatures}

\item{min_depth}{minimum alt depth for mutation signature analysis}
}
\description{
Writes out mutsig to somatic object
}
