% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{calculate_STAR_TPM}
\alias{calculate_STAR_TPM}
\title{Calculate TPM from feature counts}
\usage{
calculate_STAR_TPM(somatic, genome)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{genome}{genome data, used to map transcripts}
}
\description{
Uses effective length calculation and writes STAR_tpm to somatic
}
