% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{calculate_STAR_TPM_matrix}
\alias{calculate_STAR_TPM_matrix}
\title{Aggregates TPM counts from all samples into a matrix for protein-coding genes
defined in genome$ensembl}
\usage{
calculate_STAR_TPM_matrix(genome, SO, max_TPM = 1e+09)
}
\arguments{
\item{genome}{genome data, used to map transcripts}

\item{SO}{incoming list of somatic objects}

\item{max_TPM}{maximum allowed TPM value, exceed this and get set to 0}
}
\description{
Aggregates TPM counts from all samples into a matrix for protein-coding genes
defined in genome$ensembl
}
