% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{calculate_strelka_properties}
\alias{calculate_strelka_properties}
\title{Calculate number of strelka calls where the alt allele has at least min_depth}
\usage{
calculate_strelka_properties(somatic, min_depth = 5)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{min_depth}{minimum depth to count a call}
}
\description{
Calculate number of strelka calls where the alt allele has at least min_depth
}
