% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{compute_ploidy_with_retries}
\alias{compute_ploidy_with_retries}
\title{Use CLONET to calculate ploidy}
\usage{
compute_ploidy_with_retries(
  beta_table,
  n_cores = 1,
  library_type,
  max_homo_dels_fraction = 0.01,
  min_required_snps = 5
)
}
\arguments{
\item{beta_table}{calculated by CLONET}

\item{n_cores}{threads, default 1}

\item{library_type}{WGS or WES}

\item{max_homo_dels_fraction}{maximum allowed fraction of segment with homozygous deletion, default 0.01}

\item{min_required_snps}{minimum required SNPs within a segment, default 5}
}
\value{
the ploidy table calculated by CLONET
}
\description{
Use CLONET to calculate ploidy
}
