% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert.genome.to.scaled.x}
\alias{convert.genome.to.scaled.x}
\title{Convert chromosomes and positions to a scaled proportionate value}
\usage{
convert.genome.to.scaled.x(
  chr.list,
  loc.list,
  x.max = 1000,
  chrom_lengths = NULL
)
}
\arguments{
\item{chr.list}{chromosomes}

\item{loc.list}{locations on those chromosomes}

\item{x.max}{highest value for scaling}

\item{chrom_lengths}{length of each chromosome, defaults to hardcoded hg38 lengths}
}
\description{
Convert chromosomes and positions to a scaled proportionate value
}
