% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{copynumber_range}
\alias{copynumber_range}
\title{Extract rdbins within a genomic range}
\usage{
copynumber_range(genome, somatic, chrom, loc_start, loc_end)
}
\arguments{
\item{genome}{Genome object}

\item{somatic}{somatic object from which to extract}

\item{chrom}{chromosome}

\item{loc_start}{location to start}

\item{loc_end}{location to end}
}
\value{
subset of rdbins
}
\description{
Extract rdbins within a genomic range
}
