% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{file_check}
\alias{file_check}
\title{wrapper to check if file exists}
\usage{
file_check(func_name, fn)
}
\arguments{
\item{func_name}{report back which function was called}

\item{fn}{file path to check}
}
\description{
wrapper to check if file exists
}
