% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.split.col}
\alias{get.split.col}
\title{splits elemts of a vector}
\usage{
get.split.col(v, string, col = 0, last = F, first = F)
}
\arguments{
\item{v}{vector}

\item{string}{search string}

\item{col}{index of the token to return}

\item{last}{return the last token}

\item{first}{return the first token}
}
\description{
splits elemts of a vector
}
