% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{identify_microhomology}
\alias{identify_microhomology}
\title{Identify microhomology between two loci in the genome on the same chromosome}
\usage{
identify_microhomology(chrom, pos_start, pos_end, n_shoulder)
}
\arguments{
\item{chrom}{chromosome}

\item{pos_start}{starting locus}

\item{pos_end}{ending locus}

\item{n_shoulder}{number of nucleotides in either direction to scan}
}
\description{
Identify microhomology between two loci in the genome on the same chromosome
}
