% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{match.idx}
\alias{match.idx}
\title{return dataframe of indices into A and B restricted to perfect matches}
\usage{
match.idx(A, B, allow.multiple.B = FALSE)
}
\arguments{
\item{A}{first vector}

\item{B}{second vector}

\item{allow.multiple.B}{if TRUE, allow single hits in A to match multiple hits in B}
}
\description{
If there is more than one match between A and B, reports only the first match
}
