% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_fusions.R
\name{parse_gridss_vcf}
\alias{parse_gridss_vcf}
\title{Parse the VCF file from GRIDSS}
\usage{
parse_gridss_vcf(vcf_f)
}
\arguments{
\item{vcf_f}{vcf object to parse}
}
\description{
Parse the VCF file from GRIDSS
}
