% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{parse_manta_PR_SR}
\alias{parse_manta_PR_SR}
\title{Parse a Manta VCF file to calculate read depth in the WT and alternate}
\usage{
parse_manta_PR_SR(vcf_f)
}
\arguments{
\item{vcf_f}{VCF file read by VariantAnnotation package}
}
\description{
Parse a Manta VCF file to calculate read depth in the WT and alternate
}
