% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{plot_circlize}
\alias{plot_circlize}
\title{Generate circos plot of structural and copy number alterations}
\usage{
plot_circlize(
  SO,
  sample_id,
  chroms,
  xlims,
  ylims,
  min_SV_coverage = 10,
  horizontal_axis_lines_at,
  sv_manual = data.frame(),
  main = NA
)
}
\arguments{
\item{SO}{list of somatic objects}

\item{sample_id}{sample to plot, must be name of element in SO}

\item{chroms}{vector of chromosomes to plot}

\item{xlims}{x limits on chromosomes; 2 element vector if 1, matrix if >1 chromosome}

\item{ylims}{common y limits, 2 element vector}

\item{min_SV_coverage}{restrict SV to those with combined ALT at least this depth}

\item{horizontal_axis_lines_at}{where to draw the horizontal grid lines}

\item{sv_manual}{data frame of additional SV to plot}

\item{main}{text to draw at center, NA if leave blank}
}
\description{
Generate circos plot of structural and copy number alterations
}
