% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{plot_CNA_add_segments}
\alias{plot_CNA_add_segments}
\title{Plot a track of CNA data, called after plot_CNA_genome_background()}
\usage{
plot_CNA_add_segments(chrom_ints, starts, ends, values, color = "black")
}
\arguments{
\item{chrom_ints}{chromosomes expressed as integers, so X = 23 and chr5 = 5}

\item{starts}{positions to start segments}

\item{ends}{positions to end segments}

\item{values}{heights}

\item{color}{for line}
}
\description{
Can be called multiple times
}
