% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{plot_CNA_bins_across_samples}
\alias{plot_CNA_bins_across_samples}
\title{Plot a uncorrected bins of copy number data in a genomic region}
\usage{
plot_CNA_bins_across_samples(
  genome,
  SO,
  chrom,
  loc_start,
  loc_end,
  symbols,
  ylim
)
}
\arguments{
\item{genome}{genome object}

\item{SO}{list of somatic objects}

\item{chrom}{chromosome to start segments}

\item{loc_start}{position to start bins}

\item{loc_end}{positions to end bins}

\item{symbols}{symbols to mark with rectangles at the bottom of the plot}

\item{ylim}{ylim values for plot}
}
\value{
data frame of bins plotted
}
\description{
Plot a uncorrected bins of copy number data in a genomic region
}
