% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{plot_CNA_genome_background}
\alias{plot_CNA_genome_background}
\title{without putting any data up, create the background for a genome-wide plot of CNA data}
\usage{
plot_CNA_genome_background(genome, ylim = c(0, 5), main = "", ylab = "copies")
}
\arguments{
\item{genome}{Genome object}

\item{ylim}{ylim for plot}

\item{main}{title}

\item{ylab}{y axis label}
}
\description{
without putting any data up, create the background for a genome-wide plot of CNA data
}
