% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{plot_SNV_frequency}
\alias{plot_SNV_frequency}
\title{Plot a log10 barplot of SNV frequency}
\usage{
plot_SNV_frequency(
  SO,
  ylim = c(0, 5),
  x_labels = NULL,
  denominator = 3000,
  min_alt_count = 20
)
}
\arguments{
\item{SO}{list of somatic objects}

\item{ylim}{ylim}

\item{x_labels}{labels for objects on X axis}

\item{denominator}{divide counts by this value, 3000 would give muts / Mb}

\item{min_alt_count}{minimum reads for alternate allele to count}
}
\description{
Plot a log10 barplot of SNV frequency
}
