% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{plot_symbol_summary_grid}
\alias{plot_symbol_summary_grid}
\title{Generate a symbol by sample grid illustrating somatic alterations}
\usage{
plot_symbol_summary_grid(
  SO,
  genome,
  symbol_list,
  sample_ids,
  block.height = 6,
  block.width = 10,
  cex.y = 1,
  show_x_axis = TRUE,
  show_y_labels = TRUE,
  sex = "M",
  CNA_amp_threshold = 4,
  CNA_deepdel_threshold = 0.5,
  missense_whitelist = NA,
  germline_whitelist = NA,
  LOH_list = NA
)
}
\arguments{
\item{SO}{list of somatic objects}

\item{genome}{genome metadata generated by Genome() function}

\item{symbol_list}{vector of symbols (in genome) to plot after a call to symbol_summary}

\item{sample_ids}{names of elements in SO to plot}

\item{block.height}{grid element height}

\item{block.width}{grid element width}

\item{cex.y}{cex Y axis}

\item{show_x_axis}{show the X axis}

\item{show_y_labels}{show the Y labels}

\item{sex}{one of {M,F}, defaults to M}

\item{CNA_amp_threshold}{threshold to call amplification, default 4}

\item{CNA_deepdel_threshold}{threshold to call deep deletion, default 0.5}

\item{missense_whitelist}{only show missense mutations from symbols on this list}

\item{germline_whitelist}{only show germline alterations from symbols on this list}

\item{LOH_list}{dataframe with columns symbol,sample_id. Show LOH for these specific cases and no others.}
}
\description{
Generate a symbol by sample grid illustrating somatic alterations
}
