% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{read_bamqc_summary}
\alias{read_bamqc_summary}
\title{parses the BamQC summmary file to obtain n_mapped_paired, n_duplicated, median_insert_size, mean_mapping_quality}
\usage{
read_bamqc_summary(fn)
}
\arguments{
\item{fn}{path to file written by BamQC}
}
\description{
parses the BamQC summmary file to obtain n_mapped_paired, n_duplicated, median_insert_size, mean_mapping_quality
}
