% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{read_CNA_rdbins_cnvkit}
\alias{read_CNA_rdbins_cnvkit}
\title{Read in binned copy number data created by CNVkit}
\usage{
read_CNA_rdbins_cnvkit(fn_path, sex = "M")
}
\arguments{
\item{fn_path}{path to binned data}

\item{sex}{M or F, used to adjust copies for diploid status}
}
\description{
Output is chrom, start, end, copies
Infers interval from the gaps between bins.
}
