% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{read_CNA_rdbins_copycat}
\alias{read_CNA_rdbins_copycat}
\title{Read in binned copy number data created by copycat}
\usage{
read_CNA_rdbins_copycat(fn_path)
}
\arguments{
\item{fn_path}{path to binned data}
}
\description{
Output is chrom, start, end, copies
Infers interval from the gaps between bins.
Note that copies doesn't know whether a chromosome is diploid or not.
}
