% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{read_CNA_segments_cnvkit}
\alias{read_CNA_segments_cnvkit}
\title{Read in a segment file created by CNVkit}
\usage{
read_CNA_segments_cnvkit(fn_path, sex = "M")
}
\arguments{
\item{fn_path}{path to segment file}

\item{sex}{M or F, used to calculate copies}
}
\description{
Output has chrom, start, end, log2, depth, weight, copies, chrom_int
where chrom_in is the integer version of chromosomes (X = 23)
and copies is converted from log2 using knowledge of whether sample is M or F
}
