% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{read_CNA_segments_copycat}
\alias{read_CNA_segments_copycat}
\title{Read in a segment file created by CopyCat}
\usage{
read_CNA_segments_copycat(fn_path)
}
\arguments{
\item{fn_path}{path to segment file}
}
\description{
Output has chrom, start, end, segments, copies, chrom_int
where chrom_in is the integer version of chromosomes (X = 23)
}
