% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{read_first_token_from_file}
\alias{read_first_token_from_file}
\title{used to taste files to see what kind they are}
\usage{
read_first_token_from_file(fn)
}
\arguments{
\item{fn}{path to file to test}
}
\value{
first token in file, converted to numeric
}
\description{
used to taste files to see what kind they are
}
