% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{read_germline_pathogenic}
\alias{read_germline_pathogenic}
\title{Wrapper function to add annotated germline calls from HaplotypeCaller (GATK4)}
\usage{
read_germline_pathogenic(somatic, fn_germline)
}
\arguments{
\item{somatic}{incoming somatic object to modify}

\item{fn_germline}{germline variant file to read}
}
\description{
Writes out mutsig to somatic object
}
