% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{read_Picard_alignmentsummary}
\alias{read_Picard_alignmentsummary}
\title{parses the Picard alignment summary to obtain the number of paired reads}
\usage{
read_Picard_alignmentsummary(fn)
}
\arguments{
\item{fn}{path to file written by Picard}
}
\description{
parses the Picard alignment summary to obtain the number of paired reads
}
