% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{read_SNV_gene_summary_strelka}
\alias{read_SNV_gene_summary_strelka}
\title{Parse Strelka output from SNV and Indel text files}
\usage{
read_SNV_gene_summary_strelka(fn_strelka_clinvar_snv, fn_strelka_clinvar_indel)
}
\arguments{
\item{fn_strelka_clinvar_snv}{path to SNV file}

\item{fn_strelka_clinvar_indel}{path to indel file}
}
\description{
Parse Strelka output from SNV and Indel text files
}
