% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{read_STAR_counts}
\alias{read_STAR_counts}
\title{parse the output of a STAR mapping call to extract transcript counts and metadata of alignment
writes N_unmapped, N_multimapping, N_noFeature, N_ambiguous, and STAR_counts}
\usage{
read_STAR_counts(fn_path, somatic, genome)
}
\arguments{
\item{fn_path}{path to STAR output}

\item{somatic}{incoming somatic object to modify}

\item{genome}{genome data, used to map transcripts}
}
\description{
parse the output of a STAR mapping call to extract transcript counts and metadata of alignment
writes N_unmapped, N_multimapping, N_noFeature, N_ambiguous, and STAR_counts
}
