% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{read_STAR_mean_insert_length}
\alias{read_STAR_mean_insert_length}
\title{Parse the output of a STAR metadata output to extract mean insert length}
\usage{
read_STAR_mean_insert_length(fn_path, somatic)
}
\arguments{
\item{fn_path}{path to star metadata}

\item{somatic}{incoming somatic object to modify}
}
\description{
input file has loose text format. Writes STAR_mean_insert_length.
}
