% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{read_VCF}
\alias{read_VCF}
\title{Wrapper around readVcf}
\usage{
read_VCF(fn_vcf, build = "hg38")
}
\arguments{
\item{fn_vcf}{path to VCF file}

\item{build}{default to hg38}
}
\description{
Wrapper around readVcf
}
