% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{relative_CN}
\alias{relative_CN}
\title{Find copy number of two sets of CNA segments, e.g. pre vs post}
\usage{
relative_CN(sample_id, segments_pre, segments_post)
}
\arguments{
\item{sample_id}{sample ID to write into output dataframe}

\item{segments_pre}{segment file (e.g. from a somatic object)}

\item{segments_post}{segment file (e.g. from a somatic object)}
}
\value{
A data frame with copy number of overlapping segments in pre and post-event samples
}
\description{
This function would be used to generate a plot showing copy number across the
genome at two samples from the same individual obtained at two different timepoints.
It uses GenomicRanges to find the segments that are in both pre and post
and writes the value of copies_corr into these segments
}
