% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{Somatic}
\alias{Somatic}
\title{Constuctor for object that holds somatic data for a sample}
\usage{
Somatic(sample_base, sample_id, library_type)
}
\arguments{
\item{sample_base}{donor identifier, might not be unique}

\item{sample_id}{unique sample identifier}

\item{library_type}{One of {WES,WGS}}
}
\value{
A new Somatic object
}
\description{
Constuctor for object that holds somatic data for a sample
}
