% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{somatic_add_CNA_by_gene_from_segments}
\alias{somatic_add_CNA_by_gene_from_segments}
\title{calculate gene level copy number calls using segments defined by PURPLE and gene locations}
\usage{
somatic_add_CNA_by_gene_from_segments(somatic, genome)
}
\arguments{
\item{somatic}{somatic data object to modify}

\item{genome}{genome object created by Genome(), contains gene locations}
}
\description{
calculate gene level copy number calls using segments defined by PURPLE and gene locations
}
