% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{somatic_add_CNA_by_gene_purple}
\alias{somatic_add_CNA_by_gene_purple}
\title{Wrapper function to add gene level copy number calls from PURPLE}
\usage{
somatic_add_CNA_by_gene_purple(somatic, fn_cna)
}
\arguments{
\item{somatic}{somatic data object to modify}

\item{fn_cna}{path to gene level copy number file, required}
}
\description{
Wrapper function to add gene level copy number calls from PURPLE
}
