% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA.R
\name{somatic_add_CNA_purple}
\alias{somatic_add_CNA_purple}
\title{Wrapper function to add CNA segments from PURPLE}
\usage{
somatic_add_CNA_purple(somatic, fn_segs, fn_purity = NULL, fn_bins = NULL)
}
\arguments{
\item{somatic}{somatic data object to modify}

\item{fn_segs}{path to segment file, required}

\item{fn_purity}{path to purity summary, can be null}

\item{fn_bins}{path to binned copy numbers, can be null}
}
\description{
Wrapper function to add CNA segments from PURPLE
}
