% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_fusions.R
\name{somatic_add_fusion_linx}
\alias{somatic_add_fusion_linx}
\title{add fusions calculated by linx}
\usage{
somatic_add_fusion_linx(somatic, fn_fusions)
}
\arguments{
\item{somatic}{somatic data object to modify}

\item{fn_fusions}{file with fusions to add}
}
\description{
add fusions calculated by linx
}
