% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_fusions.R
\name{somatic_add_links_linx}
\alias{somatic_add_links_linx}
\title{Add links identified by linx}
\usage{
somatic_add_links_linx(somatic, fn_links)
}
\arguments{
\item{somatic}{somatic data object to modify}
}
\description{
Add links identified by linx
}
