% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppy.R
\name{symbol_summary}
\alias{symbol_summary}
\title{Completely characterize one gene in one sample}
\usage{
symbol_summary(
  S,
  genome,
  symbol,
  sex = "M",
  CNA_max_deepdel = 0.7,
  CNA_max_shallowdel = 1.25,
  CNA_min_amp = 3,
  SNV_method = "strelka_mutect",
  effects_missense = c("missense_variant", "missense_variant&splice_region_variant"),
  effects_inactivate = c("stop_gained", "stop_gained&splice_region_variant",
    "frameshift_variant"),
  effects_splice = c("splice_acceptor_variant&intron_variant",
    "splice_donor_variant&intron_variant")
)
}
\arguments{
\item{S}{somatic object, a list of data measurements such as SNVs and CNA data}

\item{genome}{genome metadata generated by Genome() function}

\item{symbol}{gene symbol to analyze}

\item{sex}{M or F}

\item{CNA_max_deepdel}{threshold for calling deep deltion}

\item{CNA_max_shallowdel}{threshold for calling shallow deletion}

\item{CNA_min_amp}{threshold for calling amplification}

\item{SNV_method}{which SNV summary to use, one of {strelka,strelka_mutect,mutect}}

\item{effects_missense}{list of effects to count as missense, defaults "missense_variant", "missense_variant&splice_region_variant"}

\item{effects_inactivate}{list of effects to count as inactivating, defaults "stop_gained", "stop_gained&splice_region_variant", "frameshift_variant"}

\item{effects_splice}{list of effects to count as splice inactivating, defaults "splice_acceptor_variant&intron_variant", "splice_donor_variant&intron_variant"}
}
\description{
uses any information that has been loaded into SO to characterize a symbol in one sample
for copy alterations, SNVs, inactivating SVs, and number of predicted functional alleles
}
